Move files into new folder structure
