+++ qcodes/tests/dataset/test_descriptions.py
@@ -2,9 +2,10 @@ import json
 
 import pytest
 
-from qcodes.dataset.descriptions import RunDescriber
+from qcodes.dataset.descriptions.rundescriber import RunDescriber
 from qcodes.utils.helpers import YAML
-from qcodes.dataset.dependencies import InterDependencies_, new_to_old
+from qcodes.dataset.descriptions.dependencies import InterDependencies_, \
+    new_to_old
 
 # pylint: disable=unused-import
 from qcodes.tests.dataset.interdeps_fixtures import (
